Refactor file structure for clarity & simplicity.
