+++ server/src/routes/auth/index.ts
@@ -2,13 +2,11 @@ import { Router } from "express";
 import {
   authDiscordRedirectController,
   getAuthenticatedUserController,
-  // revokeAccessTokenController,
-} from "../../controllers/auth";
+} from "../../controllers";
 
 const router = Router();
 
 router.get("/discord/redirect", authDiscordRedirectController);
 router.get("/user", getAuthenticatedUserController);
-// router.get("/revoke", revokeAccessTokenController);
 
 export default router;
